Implement key modifiers
Key modifiers basically are the keys in the keyboard. Just like
in this example `v-on:keyup.enter` once enter key was released
the function alertMe is called.

Key modifiers can also be chained for example, `v-on:keyup.enter.
space`, whenever enter key or space key is released the function
alertMe is called.